Interfaz busqueda terminada. Criterios busqueda ahora tiene constructor que te ahorra tener cargarcriterios como método
